separated our state and tabButton section into Examples.jsx
